Add support for videos in posts and pages
Adding a custom include script for videos.

Setting video width to 700px inside the twitter template
and enforcing a 16:9 aspect ratio.

Using fitvids.js to dynamically adjust the video size
when the page is viewed on small screens.

Now all you need to do is use a `youtube` or `vimeo` tag in the header
of your page or post, to embed the video at the top of the page.

Please note:

- only one video can be embedded per page / post
- Since we are using inclusion rather than a jekyll plugin this approach
  can also be used for pages hosted at github where plugins are disabled
  by default